Create README.md
Ecotrade is a global financial and digital asset community, with the goal of creating a convenient and secure decentralized trading environment, connecting users and businesses around the world. With the mission of building an advanced and reliable ecosystem, Ecotrade provides services such as financial trading, digital banking, smart financial management and effective investment solutions.